"Got everything running again!"

+++ Eclipse_Workspace/Tower Defense/src/objects/Shot.java
@@ -2,24 +2,24 @@ package objects;
 
 public class Shot {
 	
-	private int strength;		//strength of shots
+	private Double strength;		//strength of shots
 	private Double speed;			//number of shots in 'time'
 	private Double velocity;		//velocity of shots
 	
 	
-	public Shot(int pStrength, Double pSpeed, Double pVelocity)
+	public Shot(Double pStrength, Double pSpeed, Double pVelocity)
 	{
 		this.strength=pStrength;
 		this.speed=pSpeed;
 		this.velocity=pVelocity;
 	}
 	
-	public int getStrength()				//returns value of 'strength'
+	public Double getStrength()				//returns value of 'strength'
 	{
 		return strength;
 	}
 	
-	public void setStrength(int strength)	
+	public void setStrength(Double strength)	
 	{
 		this.strength=strength;
 	}
